+++ packages/tswebextension/src/lib/mv2/content-script/cookie-controller.ts
@@ -1,3 +1,5 @@
+import { CookieRule } from '../../common/content-script';
+
 interface OnRuleAppliedData {
     cookieName: string,
     cookieValue: string,
@@ -9,15 +11,6 @@ interface OnRuleAppliedData {
 
 interface OnRuleAppliedCallback {
     (data: OnRuleAppliedData): void;
-}
-
-// TODO check if this type is not duplicated
-interface Rule {
-    ruleText: string;
-    match: string;
-    isThirdParty: boolean;
-    filterId: number;
-    isAllowlist: boolean;
 }
 
 /**
@@ -58,7 +51,7 @@ export class CookieController {
      *
      * @param rules Rules to apply.
      */
-    public apply(rules: Rule[]): void {
+    public apply(rules: CookieRule[]): void {
         this.applyRules(rules);
 
         let lastCookie = document.cookie;
@@ -125,13 +118,7 @@ export class CookieController {
      * @see {@link https://github.com/AdguardTeam/Scriptlets/blob/master/src/scriptlets/remove-cookie.js}
      */
     private applyRules(
-        rules: {
-            ruleText: string;
-            match: string;
-            isThirdParty: boolean;
-            filterId: number;
-            isAllowlist: boolean;
-        }[],
+        rules: CookieRule[],
     ): void {
         document.cookie.split(';').forEach((cookieStr) => {
             const pos = cookieStr.indexOf('=');
@@ -179,7 +166,7 @@ export class CookieController {
      * @param cookieValue Cookie value.
      */
     private applyRule(
-        rule: Rule,
+        rule: CookieRule,
         cookieName: string,
         cookieValue: string,
     ): void {