+++ packages/tswebextension/CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
-## Unreleased
+## [0.3.11] - 2023-08-21
+
+### Changed
+- Updated `@adguard/tsurlfilter` to `v2.1.10`.
+- Updated `@adguard/scriptlets` to `v1.9.70`.
 
 ### Fixed
 - Applying $cookie rules on the content-script side.